wandr-backend/requirements.txt
# Core Flask Dependencies
Flask==3.1.0
Flask-CORS==5.0.0

# Environment Variables Management
python-decouple==3.8

# Development Dependencies
python-dotenv==1.0.1

# Future Dependencies (commented for now, will uncomment as needed)
# Database
# Flask-SQLAlchemy==3.1.1
# psycopg2-binary==2.9.9

# WebSocket Support
# Flask-SocketIO==5.3.6

# API Integrations (for future multi-agent system)
# requests==2.31.0
# google-generativeai==0.8.3
# twilio==9.2.4

# Agent Framework (for future)
# langchain==0.2.16
# langchain-community==0.2.16

# Async Support
# redis==5.0.8
# celery==5.3.4
